prepare v4.0 for herelink
