--- src/FlightDisplay/VideoManager.cc
@@ -1,294 +1,0 @@
-/****************************************************************************
- *
- *   (c) 2009-2016 QGROUNDCONTROL PROJECT <http://www.qgroundcontrol.org>
- *
- * QGroundControl is licensed according to the terms in the file
- * COPYING.md in the root of the source code directory.
- *
- ****************************************************************************/
-
-
-#include <QQmlContext>
-#include <QQmlEngine>
-#include <QSettings>
-#include <QUrl>
-#include <QDir>
-
-#ifndef QGC_DISABLE_UVC
-#include <QCameraInfo>
-#endif
-
-#include <VideoItem.h>
-
-#include "ScreenToolsController.h"
-#include "VideoManager.h"
-#include "QGCToolbox.h"
-#include "QGCCorePlugin.h"
-#include "QGCOptions.h"
-#include "MultiVehicleManager.h"
-#include "Settings/SettingsManager.h"
-#include "Vehicle.h"
-#include "QGCCameraManager.h"
-
-QGC_LOGGING_CATEGORY(VideoManagerLog, "VideoManagerLog")
-
-//-----------------------------------------------------------------------------
-VideoManager::VideoManager(QGCApplication* app, QGCToolbox* toolbox)
-    : QGCTool(app, toolbox)
-{
-}
-
-//-----------------------------------------------------------------------------
-VideoManager::~VideoManager()
-{
-    if(_videoReceiver) {
-        delete _videoReceiver;
-    }
-}
-
-//-----------------------------------------------------------------------------
-void
-VideoManager::setToolbox(QGCToolbox *toolbox)
-{
-   QGCTool::setToolbox(toolbox);
-   QQmlEngine::setObjectOwnership(this, QQmlEngine::CppOwnership);
-   qmlRegisterUncreatableType<VideoManager> ("QGroundControl.VideoManager", 1, 0, "VideoManager", "Reference only");
-   qmlRegisterUncreatableType<VideoReceiver>("QGroundControl",              1, 0, "VideoReceiver","Reference only");
-   qmlRegisterUncreatableType<VideoSurface> ("QGroundControl",              1, 0, "VideoSurface", "Reference only");
-   _videoSettings = toolbox->settingsManager()->videoSettings();
-   QString videoSource = _videoSettings->videoSource()->rawValue().toString();
-   connect(_videoSettings->videoSource(),   &Fact::rawValueChanged, this, &VideoManager::_videoSourceChanged);
-   connect(_videoSettings->udpPort(),       &Fact::rawValueChanged, this, &VideoManager::_udpPortChanged);
-   connect(_videoSettings->rtspUrl(),       &Fact::rawValueChanged, this, &VideoManager::_rtspUrlChanged);
-   connect(_videoSettings->tcpUrl(),        &Fact::rawValueChanged, this, &VideoManager::_tcpUrlChanged);
-   connect(_videoSettings->aspectRatio(),   &Fact::rawValueChanged, this, &VideoManager::_aspectRatioChanged);
-
-   MultiVehicleManager *pVehicleMgr = qgcApp()->toolbox()->multiVehicleManager();
-   connect(pVehicleMgr, &MultiVehicleManager::activeVehicleChanged, this, &VideoManager::_setActiveVehicle);
-
-   _videoStreamControl = new VideoStreamControl();
-   connect(_videoStreamControl, &VideoStreamControl::videoStreamUrlChanged, this, &VideoManager::_videoStreamUrlChanged);
-
-#if defined(QGC_GST_STREAMING)
-#ifndef QGC_DISABLE_UVC
-   // If we are using a UVC camera setup the device name
-   _updateUVC();
-#endif
-
-    emit isGStreamerChanged();
-    qCDebug(VideoManagerLog) << "New Video Source:" << videoSource;
-    _videoReceiver = toolbox->corePlugin()->createVideoReceiver(this);
-    _updateSettings();
-    if(isGStreamer()) {
-        _videoReceiver->start();
-    } else {
-        _videoReceiver->stop();
-    }
-
-#endif
-}
-
-//-----------------------------------------------------------------------------
-double VideoManager::aspectRatio()
-{
-    if(_activeVehicle && _activeVehicle->dynamicCameras()) {
-        QGCVideoStreamInfo* pInfo = _activeVehicle->dynamicCameras()->currentStreamInstance();
-        if(pInfo) {
-            return pInfo->aspectRatio();
-        }
-    }
-    return _videoSettings->aspectRatio()->rawValue().toDouble();
-}
-
-//-----------------------------------------------------------------------------
-bool
-VideoManager::autoStreamConfigured()
-{
-    if(_activeVehicle && _activeVehicle->dynamicCameras()) {
-        QGCVideoStreamInfo* pInfo = _activeVehicle->dynamicCameras()->currentStreamInstance();
-        if(pInfo) {
-            return !pInfo->uri().isEmpty();
-        }
-    }
-    return false;
-}
-
-//-----------------------------------------------------------------------------
-void
-VideoManager::_updateUVC()
-{
-#ifndef QGC_DISABLE_UVC
-    QString videoSource = _videoSettings->videoSource()->rawValue().toString();
-    QList<QCameraInfo> cameras = QCameraInfo::availableCameras();
-    for (const QCameraInfo &cameraInfo: cameras) {
-        if(cameraInfo.description() == videoSource) {
-            _videoSourceID = cameraInfo.deviceName();
-            emit videoSourceIDChanged();
-            qCDebug(VideoManagerLog) << "Found USB source:" << _videoSourceID << " Name:" << videoSource;
-            break;
-        }
-    }
-#endif
-}
-
-//-----------------------------------------------------------------------------
-void
-VideoManager::_videoSourceChanged()
-{
-    _updateUVC();
-    emit hasVideoChanged();
-    emit isGStreamerChanged();
-    emit isAutoStreamChanged();
-    _restartVideo();
-}
-
-//-----------------------------------------------------------------------------
-void
-VideoManager::_udpPortChanged()
-{
-    _restartVideo();
-}
-
-//-----------------------------------------------------------------------------
-void
-VideoManager::_rtspUrlChanged()
-{
-    _restartVideo();
-}
-
-void VideoManager::_videoStreamUrlChanged(void)
-{
-    _restartVideo();
-}
-
-//-----------------------------------------------------------------------------
-void
-VideoManager::_tcpUrlChanged()
-{
-    _restartVideo();
-}
-
-//-----------------------------------------------------------------------------
-bool
-VideoManager::hasVideo()
-{
-    if(autoStreamConfigured()) {
-        return true;
-    }
-    QString videoSource = _videoSettings->videoSource()->rawValue().toString();
-    return !videoSource.isEmpty() && videoSource != VideoSettings::videoSourceNoVideo && videoSource != VideoSettings::videoDisabled;
-}
-
-//-----------------------------------------------------------------------------
-bool
-VideoManager::isGStreamer()
-{
-#if defined(QGC_GST_STREAMING)
-    QString videoSource = _videoSettings->videoSource()->rawValue().toString();
-    return
-        videoSource == VideoSettings::videoSourceUDP ||
-        videoSource == VideoSettings::videoSourceRTSP ||
-        videoSource == VideoSettings::videoSourceTCP ||
-        videoSource == VideoSettings::videoSourceMPEGTS ||
-        videoSource == VideoSettings::videoSourceAuto ||
-        autoStreamConfigured();
-#else
-    return false;
-#endif
-}
-
-//-----------------------------------------------------------------------------
-#ifndef QGC_DISABLE_UVC
-bool
-VideoManager::uvcEnabled()
-{
-    return QCameraInfo::availableCameras().count() > 0;
-}
-#endif
-
-//-----------------------------------------------------------------------------
-void
-VideoManager::_updateSettings()
-{
-    if(!_videoSettings || !_videoReceiver)
-        return;
-    //-- Auto discovery
-    if(_activeVehicle && _activeVehicle->dynamicCameras()) {
-        QGCVideoStreamInfo* pInfo = _activeVehicle->dynamicCameras()->currentStreamInstance();
-        if(pInfo) {
-            switch(pInfo->type()) {
-                case VIDEO_STREAM_TYPE_RTSP:
-                case VIDEO_STREAM_TYPE_TCP_MPEG:
-                    _videoReceiver->setUri(pInfo->uri());
-                    break;
-                case VIDEO_STREAM_TYPE_RTPUDP:
-                    _videoReceiver->setUri(QStringLiteral("udp://0.0.0.0:%1").arg(pInfo->uri()));
-                    break;
-                case VIDEO_STREAM_TYPE_MPEG_TS_H264:
-                    _videoReceiver->setUri(QStringLiteral("mpegts://0.0.0.0:%1").arg(pInfo->uri()));
-                    break;
-                default:
-                    _videoReceiver->setUri(pInfo->uri());
-                    break;
-            }
-            return;
-        }
-    }
-    QString source = _videoSettings->videoSource()->rawValue().toString();
-    if (source == VideoSettings::videoSourceUDP)
-        _videoReceiver->setUri(QStringLiteral("udp://0.0.0.0:%1").arg(_videoSettings->udpPort()->rawValue().toInt()));
-    else if (source == VideoSettings::videoSourceRTSP)
-        _videoReceiver->setUri(_videoSettings->rtspUrl()->rawValue().toString());
-    else if (source == VideoSettings::videoSourceTCP)
-        _videoReceiver->setUri(QStringLiteral("tcp://%1").arg(_videoSettings->tcpUrl()->rawValue().toString()));
-    else if (_videoSettings->videoSource()->rawValue().toString() == VideoSettings::videoSourceAuto)
-        _videoReceiver->setUri(_videoStreamControl->videoStreamUrl());
-}
-
-//-----------------------------------------------------------------------------
-void
-VideoManager::_restartVideo()
-{
-#if defined(QGC_GST_STREAMING)
-    qCDebug(VideoManagerLog) << "Restart video streaming";
-    if(!_videoReceiver)
-        return;
-    _videoReceiver->stop();
-    _updateSettings();
-    _videoReceiver->start();
-#endif
-}
-
-//----------------------------------------------------------------------------------------
-void
-VideoManager::_setActiveVehicle(Vehicle* vehicle)
-{
-    if(_activeVehicle) {
-        if(_activeVehicle->dynamicCameras()) {
-            QGCCameraControl* pCamera = _activeVehicle->dynamicCameras()->currentCameraInstance();
-            if(pCamera) {
-                pCamera->stopStream();
-            }
-            disconnect(_activeVehicle->dynamicCameras(), &QGCCameraManager::streamChanged, this, &VideoManager::_restartVideo);
-        }
-    }
-    _activeVehicle = vehicle;
-    if(_activeVehicle) {
-        if(_activeVehicle->dynamicCameras()) {
-            connect(_activeVehicle->dynamicCameras(), &QGCCameraManager::streamChanged, this, &VideoManager::_restartVideo);
-            QGCCameraControl* pCamera = _activeVehicle->dynamicCameras()->currentCameraInstance();
-            if(pCamera) {
-                pCamera->resumeStream();
-            }
-        }
-    }
-    emit autoStreamConfiguredChanged();
-    _restartVideo();
-}
-
-//----------------------------------------------------------------------------------------
-void
-VideoManager::_aspectRatioChanged()
-{
-    emit aspectRatioChanged();
-}
