--- src/FlightDisplay/VideoManager.h
@@ -1,111 +1,0 @@
-/****************************************************************************
- *
- *   (c) 2009-2016 QGROUNDCONTROL PROJECT <http://www.qgroundcontrol.org>
- *
- * QGroundControl is licensed according to the terms in the file
- * COPYING.md in the root of the source code directory.
- *
- ****************************************************************************/
-
-
-#ifndef VideoManager_H
-#define VideoManager_H
-
-#include <QObject>
-#include <QTimer>
-#include <QTime>
-#include <QUrl>
-
-#include "QGCMAVLink.h"
-#include "QGCLoggingCategory.h"
-#include "VideoReceiver.h"
-#include "QGCToolbox.h"
-#include "VideoStreamControl.h"
-
-Q_DECLARE_LOGGING_CATEGORY(VideoManagerLog)
-
-class VideoSettings;
-class Vehicle;
-class Joystick;
-
-class VideoManager : public QGCTool
-{
-    Q_OBJECT
-
-public:
-    VideoManager    (QGCApplication* app, QGCToolbox* toolbox);
-    ~VideoManager   ();
-
-    Q_PROPERTY(bool             hasVideo                READ    hasVideo                                    NOTIFY hasVideoChanged)
-    Q_PROPERTY(bool             isGStreamer             READ    isGStreamer                                 NOTIFY isGStreamerChanged)
-    Q_PROPERTY(bool             isTaisync               READ    isTaisync       WRITE   setIsTaisync        NOTIFY isTaisyncChanged)
-    Q_PROPERTY(QString          videoSourceID           READ    videoSourceID                               NOTIFY videoSourceIDChanged)
-    Q_PROPERTY(bool             uvcEnabled              READ    uvcEnabled                                  CONSTANT)
-    Q_PROPERTY(bool             fullScreen              READ    fullScreen      WRITE   setfullScreen       NOTIFY fullScreenChanged)
-    Q_PROPERTY(VideoReceiver*   videoReceiver           READ    videoReceiver                               CONSTANT)
-    Q_PROPERTY(double           aspectRatio             READ    aspectRatio                                 NOTIFY aspectRatioChanged)
-    Q_PROPERTY(bool             autoStreamConfigured    READ    autoStreamConfigured                        NOTIFY autoStreamConfiguredChanged)
-    Q_PROPERTY(VideoStreamControl* videoStreamControl READ  videoStreamControl                          CONSTANT)
-
-    bool        hasVideo            ();
-    bool        isGStreamer         ();
-    bool        isAutoStream        ();
-    bool        isTaisync           () { return _isTaisync; }
-    bool        fullScreen          () { return _fullScreen; }
-    QString     videoSourceID       () { return _videoSourceID; }
-    double      aspectRatio         ();
-    bool        autoStreamConfigured();
-
-    VideoReceiver*  videoReceiver   () { return _videoReceiver; }
-    VideoStreamControl* videoStreamControl () { return _videoStreamControl; }
-
-#if defined(QGC_DISABLE_UVC)
-    bool        uvcEnabled          () { return false; }
-#else
-    bool        uvcEnabled          ();
-#endif
-
-    void        setfullScreen       (bool f) { _fullScreen = f; emit fullScreenChanged(); }
-    void        setIsTaisync        (bool t) { _isTaisync = t;  emit isTaisyncChanged(); }
-
-    // Override from QGCTool
-    void        setToolbox          (QGCToolbox *toolbox);
-
-    Q_INVOKABLE void startVideo     () { _videoReceiver->start(); }
-    Q_INVOKABLE void stopVideo      () { _videoReceiver->stop();  }
-
-signals:
-    void hasVideoChanged            ();
-    void isGStreamerChanged         ();
-    void videoSourceIDChanged       ();
-    void fullScreenChanged          ();
-    void isAutoStreamChanged        ();
-    void isTaisyncChanged           ();
-    void aspectRatioChanged         ();
-    void autoStreamConfiguredChanged();
-
-private slots:
-    void _videoSourceChanged        ();
-    void _udpPortChanged            ();
-    void _rtspUrlChanged            ();
-    void _tcpUrlChanged             ();
-    void _updateUVC                 ();
-    void _setActiveVehicle          (Vehicle* vehicle);
-    void _aspectRatioChanged        ();
-    void _restartVideo              ();
-    void _videoStreamUrlChanged     ();
-
-private:
-    void _updateSettings            ();
-
-private:
-    bool            _isTaisync          = false;
-    VideoReceiver*  _videoReceiver      = nullptr;
-    VideoSettings*  _videoSettings      = nullptr;
-    QString         _videoSourceID;
-    bool            _fullScreen         = false;
-    VideoStreamControl* _videoStreamControl;
-    Vehicle*        _activeVehicle      = nullptr;
-};
-
-#endif
