+++ src/VideoStreaming/VideoStreaming.pri
@@ -64,7 +64,7 @@ LinuxBuild {
 } else:AndroidBuild {
     #- gstreamer assumed to be installed in $$PWD/../../gstreamer-1.0-android-universal-1.14.4/***
     contains(QT_ARCH, arm) {
-        GST_ROOT = $$PWD/../../gstreamer-1.0-android-universal-1.14.4/armv7
+        GST_ROOT = $$PWD/../../gstreamer/armv7
     } else:contains(QT_ARCH, arm64) {
         GST_ROOT = $$PWD/../../gstreamer-1.0-android-universal-1.14.4/arm64
     } else {
